requirements.txt
Django==1.8.2
django-debug-toolbar==1.3.0
django-filer==0.9.11
django-filter==0.10.0
django-localflavor==1.1
django-mptt==0.7.4
django-polymorphic==0.7.1
django-redis-cache==0.13.1
django-suit==0.2.13
djangorestframework==3.1.3
easy-thumbnails==2.2
flake8==2.4.1
gunicorn==19.3.0
hiredis==0.2.0
importmagic==0.1.3
ipython==3.1.0
jedi==0.9.0
Markdown==2.6.2
mccabe==0.3
oauthlib==0.7.2
pbr==1.1.0
pep8==1.5.7
Pillow==2.8.2
psycopg2==2.6
pyflakes==0.8.1
PyJWT==1.3.0
python-openid==2.2.5
python-social-auth==0.2.10
redis==2.10.3
requests==2.7.0
requests-oauthlib==0.5.0
rope==0.10.2
six==1.9.0
sqlparse==0.1.15
stevedore==1.5.0
Unidecode==0.4.17
virtualenv==13.0.3
virtualenv-clone==0.2.5
virtualenvwrapper==4.6.0
